Move `gh2gcs` functionality to `kpromo gh`
